OllamaRAGDemo/requirements.txt
chromadb==0.4.22
langchain==0.1.6
langchain-community==0.0.20
beautifulsoup4==4.12.3
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
ollama==0.1.6
sentence-transformers==2.3.1
tiktoken==0.5.2
# Enhanced document processing dependencies
pdf2image>=1.16.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
